refactor: moved GetMembers endpoint to members subdirectory
