+++ src/main/scala/api/storages/Endpoints.scala
@@ -1,6 +1,7 @@
 package api.storages
 
 import api.storages.ingredients.storagesIngredientsEndpoints
+import api.storages.members.storagesMembersEndpoints
 import domain.UserId
 
 import sttp.tapir.ztapir.*
@@ -15,5 +16,5 @@ val storageEndpoints = List(
   delete,
   getAll,
   getSummary,
-  getMembers
 ) ++ storagesIngredientsEndpoints
+  ++ storagesMembersEndpoints
